have panel manage kernel specifications
